+++ opendut-types/src/peer/mod.rs
@@ -77,8 +77,8 @@ impl PeerName {
     pub const MIN_LENGTH: usize = 4;
     pub const MAX_LENGTH: usize = 64;
 
-    pub fn value(self) -> String {
-        self.0
+    pub fn value(&self) -> &str {
+        &self.0
     }
 }